II Core, II Windows; Performance and Drawing Overhaul v0.2 (#112) II Core - Debugged type casting changes - Debugged drawing routine implementation II Windows - Converted Bitmap via PNG to BitmapImage   - Displays transparency properly - Fixed memory leaks evading garbage collection!   - Timers unsubscribed, Dispose()'ed   - Devices unsubscribed from PatientEvents, Dispose()'ed

+++ II Core/Classes/Math.cs
@@ -68,8 +68,8 @@ namespace II {
             return (current - min) / (max - min);
         }
 
-        public static int InverseLerp (int min, int max, float current) {
-            return (int)((current - min) / (max - min));
+        public static float InverseLerp (int min, int max, float current) {
+            return ((current - min) / (max - min));
         }
 
         public static double RandomDouble (double min, double max) {
